requirements.txt
# Dependências da aplicação
PyYAML==6.0.2

# Visão computacional e YOLO
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
albumentations>=1.3.0
torch>=2.0.0
torchvision>=0.15.0

# Feature selection e análise
scikit-learn>=1.3.0

# Gráficos para métricas
matplotlib>=3.7.0

